// lib/engine/types.ts
export type BusinessTypeCode = "x" | "y" | "z" | "w" | "v";
export type IncomeBand = 1 | 2 | 3 | 4 | 5;
export type FilingMode = "estimate" | "exact";
export type ResidentialStatus = "resident" | "nri" | "rnor";
export type TaxRegime = "old" | "new";

export interface SalaryInput {
  gross_salary: number;
  basic_salary: number;
  hra_received?: number;
  actual_rent_paid?: number;
  city_tier?: "metro" | "non_metro";
  professional_tax?: number;
  lta_claimed?: number;
  perquisites_taxable?: number;
  multiple_employers?: boolean;
  employer_nps_contribution?: number;
}

export interface HousePropertyInput {
  property_type?: "self_occupied" | "let_out" | "none";
  annual_rent_received?: number;
  municipal_tax?: number;
  home_loan_interest?: number;
  home_loan_principal?: number;
  pre_construction_interest?: number;
}

export interface OtherIncomeInput {
  fd_interest?: number;
  savings_account_interest?: number;
  dividend_income?: number;
}

export interface CapitalGainsInput {
  stcg_111a?: number;
  ltcg_112a?: number;
  stcg_other?: number;
  ltcg_other?: number;
  stcl_equity?: number;
  ltcl?: number;
}

export interface DeductionsInput {
  epf?: number;
  ppf?: number;
  elss?: number;
  lic_premium?: number;
  nsc?: number;
  home_loan_principal?: number;
  tuition_fees?: number;
  other_80c?: number;
  health_insurance_self?: number;
  health_insurance_parents?: number;
  parents_senior?: boolean;
  nps_self?: number;
  education_loan_interest?: number;
  rent_paid_no_hra?: number;
  donations_100pct?: number;
  donations_50pct?: number;
  savings_interest_deduction?: number;
  self_disability?: boolean;
  disability_severe?: boolean;
}

export interface TaxPaidInput {
  tds_salary?: number;
  tds_other?: number;
  advance_tax_paid?: number;
  self_assessment_tax_paid?: number;
}

export interface BusinessInput {
  business_type?:
    | "none"
    | "presumptive_business"
    | "presumptive_profession"
    | "regular_books";
  turnover?: number;
  digital_turnover_pct?: number;
  gross_professional_receipts?: number;
  actual_gross_receipts?: number;
  actual_expenses?: number;
  profession_name?: string;
  cash_receipts_pct?: number;
}

export interface ProfileFlags {
  income_band?: IncomeBand;
  business_type_code?: BusinessTypeCode;
  is_director?: boolean;
  has_unlisted_equity?: boolean;
  has_foreign_income?: boolean;
  has_foreign_assets?: boolean;
  tds_deducted_194n?: boolean;
  esop_tax_deferred?: boolean;
  agricultural_income?: number;
}

export interface DocumentFlags {
  has_form16?: boolean;
  has_ais?: boolean;
  has_form26as?: boolean;
  has_bank_interest_cert?: boolean;
  has_home_loan_cert?: boolean;
  has_capital_gains_statement?: boolean;
}

export interface UserInput {
  age: number;
  residential_status?: ResidentialStatus;
  assessment_year?: string;
  mode?: FilingMode;
  salary?: SalaryInput;
  house_property?: HousePropertyInput;
  other_income?: OtherIncomeInput;
  capital_gains?: CapitalGainsInput;
  deductions?: DeductionsInput;
  taxes_paid?: TaxPaidInput;
  business?: BusinessInput;
  profile_flags?: ProfileFlags;
  documents?: DocumentFlags;
}